Split StartEndUI into pieces and move to core-ui ..
.. so it can eventually be shared
